% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{read_xml_dwa_m150}
\alias{read_xml_dwa_m150}
\title{Read XML file coded as described in DWA-M 150}
\usage{
read_xml_dwa_m150(xml, encoding = "ISO-8859-1")
}
\arguments{
\item{xml}{path to XML file}

\item{encoding}{encoding used in the XML file, passed to
\code{\link[xml2]{read_xml}}}
}
\value{
list with elements \code{fd}, \code{hg}, \code{hi}, \code{hz}, each
 of which is a data frame containing the corresponding data as described
 in DWA-M150
}
\description{
Read XML file coded as described in DWA-M 150
}
\examples{
xml <- system.file("extdata/example.xml", package = "kwb.dwa.m150")
content <- read_xml_dwa_m150(xml)

str(content)

}
